backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
aiohttp==3.9.1
ccxt==4.1.77
pandas==2.1.4
numpy==1.25.2
ta==0.10.2
langchain==0.0.350
openai==1.3.7
stripe==7.8.0
cryptography==41.0.8
requests==2.31.0
websockets==12.0